+++ example/Main.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE LambdaCase #-}
+
 module Main
   ( main
   ) where
@@ -5,4 +7,6 @@ module Main
 import Web.Browser ( openBrowser )
 
 main :: IO ()
-main = openBrowser "https://haskell.org/" >>= print
+main = openBrowser "https://haskell.org/" >>= \case
+  True -> putStrLn "The operation succeeded!"
+  False -> putStrLn "The operation failed!"