+++ lib/Web/Browser.hs
@@ -1,30 +1,26 @@
-{-# LANGUAGE CPP #-}
-
 module Web.Browser
   ( openBrowser
   ) where
 
-#if defined(mingw32_HOST_OS)
-import Web.Browser.Windows ( openBrowserWindows )
-#else
-import Data.List ( isInfixOf )
-import System.Info ( os )
-import Web.Browser.Linux ( openBrowserLinux )
-import Web.Browser.OSX ( openBrowserOSX )
-#endif
+import qualified Web.Browser.OS as OS
 
 -- | Seeks to open the given URL in the user's preferred web browser. Returns
--- whether or not the operation succeeded. Throws an 'ErrorCall' exception if
--- the operating system is unsupported.
+-- whether or not the operation succeeded. No checks are performed on the
+-- validity of the given URL.
+--
+-- Implemented using:
+--
+-- * on Windows, the \'open\' operation provided by the Win32 API;
+--
+-- * on macOS, the \'open\' application, if it is on the user's PATH; and
+--
+-- * on Linux, FreeBSD, OpenBSD or NetBSD, the \'xdg-open\' application, if it
+--   is on the user's PATH, via \'sh\' to allow the application's output to be
+--   silenced.
+--
+-- On other operating systems, the operation always fails.
 openBrowser ::
      String
      -- ^ URL
   -> IO Bool
-#if defined(mingw32_HOST_OS)
-openBrowser = openBrowserWindows
-#else
-openBrowser
-  | any (`isInfixOf` os) ["linux", "bsd"] = openBrowserLinux
-  | "darwin" `isInfixOf` os               = openBrowserOSX
-  | otherwise                             = error "unsupported platform"
-#endif
+openBrowser = OS.openBrowser