+++ lib/unsupported-os/Web/Browser/OS.hs
@@ -1,0 +1,9 @@
+--------------------------------------------------------------------------------
+-- For unsupported operating systems
+--------------------------------------------------------------------------------
+
+module Web.Browser.OS
+  ( openBrowser
+  ) where
+
+openBrowser = const (pure False) -- Operation never succeeds